+++ src/contentIO/src/EXIParser.c
@@ -49,7 +49,7 @@
 #include "headerDecode.h"
 #include "memManagement.h"
 
-void parseEXI(char* binaryStream, uint32_t bufLen, ContentHandler* handler, void* app_data)
+void parseEXI(char* binaryStream, size_t bufLen, ContentHandler* handler, void* app_data)
 {
 	errorCode tmp_err_code = UNEXPECTED_ERROR;
 	EXIStream strm;
